fixing exercises and adding new files to cut the main file
